Basic bootloader + Qemu start script
This is just a test, the purpose is not to write our own Bootloader, we will still use syslinux.
