+++ source/Kernel-NativeAOT/Bootloader.csproj
@@ -1,0 +1,46 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+	  <OutputType>Exe</OutputType>
+	  <TargetFramework>net5.0</TargetFramework>
+	  <LangVersion>9.0</LangVersion>
+	  <EnableDefaultItems>false</EnableDefaultItems>
+	  <Platforms>x64</Platforms>
+	  <Configurations>Debug;Release</Configurations>
+	  <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+	  <Optimize Condition="'$(Configuration)'=='Release'">true</Optimize>
+
+	  <NoStdLib>true</NoStdLib>
+	  <NoConfig>true</NoConfig>
+	  <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+	  <GenerateTargetFrameworkAttribute>false</GenerateTargetFrameworkAttribute>
+	  <RuntimeMetadataVersion>v4.0.30319</RuntimeMetadataVersion>
+
+	  <IlcGenerateDgmlFile>true</IlcGenerateDgmlFile>
+	  <IlcGenerateMapFile>true</IlcGenerateMapFile>
+	  <IlcOptimizationPreference Condition="'$(Configuration)'=='Release'">Size</IlcOptimizationPreference>
+	  <IlcDisableReflection>true</IlcDisableReflection>
+	  <IlcSystemModule>Bootloader</IlcSystemModule>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<Compile Include="*.cs" />
+		<Compile Include="build.cmd" />
+	</ItemGroup>
+
+  <ItemGroup>
+    <LinkerArg Include="/subsystem:EFI_APPLICATION /entry:EfiMain" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.DotNet.ILCompiler" Version="1.0.0-alpha-*" />
+  </ItemGroup>
+
+  <Target Name="CustomizeReferences" BeforeTargets="BeforeCompile" AfterTargets="FindReferenceAssembliesForReferences">
+    <ItemGroup>
+      <ReferencePathWithRefAssemblies Remove="@(ReferencePathWithRefAssemblies)" />
+      <ReferencePath Remove="@(ReferencePath)" />
+    </ItemGroup>
+  </Target>
+
+</Project>
