+++ source/Kernel-NativeAOT/Runtime.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Runtime.InteropServices;
-
-namespace System
+﻿namespace System
 {
     public class Object { public IntPtr m_pEEType; } // The layout of object is a contract with the compiler.
     public struct Void { }
@@ -22,7 +19,7 @@ namespace System
     public abstract class ValueType { }
     public abstract class Enum : ValueType { }
     public struct Nullable<T> where T : struct { }
-    
+
     public abstract class String { public readonly int Length; }
     public abstract class Array { }
     public abstract class Delegate { }
@@ -108,78 +105,3 @@ namespace Internal.Runtime.CompilerHelpers
     }
 }
 #endregion
-
-[StructLayout(LayoutKind.Sequential)]
-public struct EFI_HANDLE
-{
-    private IntPtr _handle;
-}
-
-[StructLayout(LayoutKind.Sequential)]
-public unsafe readonly struct EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL
-{
-    private readonly IntPtr _pad;
-
-    public readonly delegate* unmanaged<void*, char*, void*> OutputString;
-}
-
-[StructLayout(LayoutKind.Sequential)]
-public readonly struct EFI_TABLE_HEADER
-{
-    public readonly ulong Signature;
-    public readonly uint Revision;
-    public readonly uint HeaderSize;
-    public readonly uint Crc32;
-    public readonly uint Reserved;
-}
-
-[StructLayout(LayoutKind.Sequential)]
-unsafe public readonly struct EFI_SYSTEM_TABLE
-{
-    public readonly EFI_TABLE_HEADER Hdr;
-    public readonly char* FirmwareVendor;
-    public readonly uint FirmwareRevision;
-    public readonly EFI_HANDLE ConsoleInHandle;
-    public readonly void* ConIn;
-    public readonly EFI_HANDLE ConsoleOutHandle;
-    public readonly EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL* ConOut;
-}
-
-public unsafe class Program
-{
-    static void Main() { }
-
-    static EFI_SYSTEM_TABLE* efitable;
-
-    [System.Runtime.RuntimeExport("EfiMain")]
-    public static long EfiMain(IntPtr imageHandle, EFI_SYSTEM_TABLE* systemTable)
-    {
-        efitable = systemTable;
-
-        Write("Hello from Cosmos! This is an EFI application compiled in C# and using NativeAOT.\r\n");
-        Write("EFI Firmware Vendor: ");
-        systemTable->ConOut->OutputString(systemTable->ConOut, systemTable->FirmwareVendor);
-        Write("\r\n");
-
-        Class emptyclass = new Class();
-        emptyclass.SayHello();
-
-        while (true);
-    }
-
-    public static void Write(string str)
-    {
-        fixed (char* pHello = str)
-        {
-            efitable->ConOut->OutputString(efitable->ConOut, pHello);
-        }
-    }
-}
-
-class Class
-{
-    public void SayHello()
-    {
-        Program.Write("Hello from a class!\r\n");
-    }
-}
